+++ Minimal/Code/Debug/DebugFunctions.py
@@ -23,6 +23,12 @@ def set_world_size(size: _float) -> None:
 
 
 # --------------------------------------------------
-def simulate(filename: _str, sim_unlocks: dict[Unlocks, _float] | Iterable[Unlocks] | Unlocks, sim_items: dict[Item, _float], sim_globals: dict[_str, Any], seed: _float, speedup: _float) -> _float:
+def simulate(
+		filename: _str,
+		sim_unlocks: dict[Unlocks, _float] | _dict[Unlocks, _float] | Iterable[Unlocks] | Unlocks,
+		sim_items: dict[Item, _float] | _dict[Item, _float],
+		sim_globals: dict[_str, Any] | _dict[_str, Any],
+		seed: _float, speedup: _float
+	) -> _float:
 	# Docstring: simulate
 	...